collector/requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.0.1
psycopg2-binary==2.9.5
requests==2.28.2
setuptools==65.6.3
soupsieve==2.3.2.post1
urllib3==1.26.14
Werkzeug==2.2.2
newspaper3k==0.2.8
openai==0.27.4
tiktoken==0.3.3
supabase==1.0.3
feedparser==6.0.10
python-dotenv==1.0.1
cloudscraper==1.2.71
lxml_html_clean==0.4.1
